Reorganize test suite with desired ORM
