+++ gemfiles/active_record_42.gemfile
@@ -2,6 +2,7 @@ source 'https://rubygems.org'
 
 # Specify your gem's dependencies in pluck_all.gemspec
 
+gem "sqlite3",      "~> 1.3.0"
 gem "activerecord", "~> 4.2.0"
 gem "carrierwave",  "~> 0.11.0"
