+++ pluck_all.gemspec
@@ -29,6 +29,5 @@ Gem::Specification.new do |spec|
 
   spec.add_development_dependency "bundler", "~> 1.11"
   spec.add_development_dependency "rake", "~> 12.0"
-  spec.add_development_dependency "sqlite3", "~> 1.3"
   spec.add_development_dependency "minitest", "~> 5.0"
 end
