+++ test/active_record/support/models/user.rb
