--- test/lib/models/mongoid.rb
@@ -1,5 +1,0 @@
-module Mongoid
-  def self.table_name_prefix
-    'mongoid_'
-  end
-end
