--- test/lib/mongoid_seeds.rb
@@ -1,5 +1,0 @@
-# frozen_string_literal: true
-Mongoid::User.delete_all
-Mongoid::User.create(name: 'Pearl Shi', age: 18)
-Mongoid::User.create(name: 'Rumble Huang', age: 20)
-Mongoid::User.create(name: 'Khiav Reoy', age: 20)
