+++ test/mongoid/mongoid_test_helper.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+require "test_helper"
+
+if defined?(Mongoid)
+  require_relative "support/config"
+  require_relative "support/seeds"
+end
