+++ test/mongoid/support/seeds.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+User.delete_all
+User.create(name: 'Pearl Shi', age: 18)
+User.create(name: 'Rumble Huang', age: 20)
+User.create(name: 'Khiav Reoy', age: 20)
