+++ test/test_helper.rb
@@ -1,15 +1,8 @@
 # frozen_string_literal: true
+$LOAD_PATH.unshift File.expand_path('../../lib', __FILE__)
+
 require "simplecov"
 SimpleCov.start
 
-$LOAD_PATH.unshift File.expand_path('../../lib', __FILE__)
-require 'pluck_all'
-
-require 'minitest/autorun'
-
-ActiveRecord::Base.establish_connection(
-  "adapter"  => "sqlite3",
-  "database" => ":memory:",
-)
-require_relative 'carrierwave_test_helper'
-require_relative 'lib/seeds'
+require "pluck_all"
+require "minitest/autorun"
